moves browserRouter to  app.jsx
