+++ src/Elements/Layout.jsx
@@ -1,0 +1,13 @@
+import { Outlet } from "react-router-dom";
+import Header from "./Header";
+
+const Layout = () => {
+  return (
+    <>
+      <Header />
+      <Outlet />
+    </>
+  );
+};
+
+export default Layout;
